ais_implement: implement basic AIS classes
AISBase class:
Implemented for finding solutions to various generative tasks (eg. set Weights for NN, find required vector of numbers etc.), where a quality of the solution suggested by AIS can be calculated.
Needs to be inherited and implementations of few methods in order to be used. For more info about how to use it and an example, see "User Manual" section in AISBase.h file and AIS_VecUint.h file as an example.

AIS_VecUint class:
Implemented as a demonstration of using AISBase class. It generates a required vector of uints, specified in the "answer" class variable, using Artificial Immune System's methods and algorithm.

* additionally: implemented file AIS_template.h. It contains functionality, same as AISBase, but in form of a namespace with functions, instead of a class.

Signed-off-by: SlavaGo <vyacheslav.golubnichiy@gmail.com>